+++ Classes/Domain/Model/Event/Event.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace OliverKlee\Seminars\Domain\Model\Event;
 
+use OliverKlee\Seminars\Domain\Model\RawDataInterface;
+use OliverKlee\Seminars\Domain\Model\RawDataTrait;
 use TYPO3\CMS\Extbase\DomainObject\AbstractEntity;
 
 /**
@@ -12,36 +14,7 @@ use TYPO3\CMS\Extbase\DomainObject\AbstractEntity;
  * Mostly, this class exists in order to get single-table inheritance working. In the code,
  * please use `EventInterface` as much as possible instead.
  */
-abstract class Event extends AbstractEntity implements EventInterface
+abstract class Event extends AbstractEntity implements RawDataInterface, EventInterface
 {
-    /**
-     * @var array<string, string|int|float|null>
-     * @TYPO3\CMS\Extbase\Annotation\ORM\Transient
-     * @internal
-     */
-    protected $rawData;
-
-    /**
-     * Returns the raw data as it is stored in the database.
-     *
-     * @return array<string, string|int|float|null>|null
-     *
-     * @internal
-     */
-    public function getRawData(): ?array
-    {
-        return $this->rawData;
-    }
-
-    /**
-     * Sets the raw data as it is stored in the database.
-     *
-     * @param array<string, string|int|float|null> $rawData
-     *
-     * @internal
-     */
-    public function setRawData(array $rawData): void
-    {
-        $this->rawData = $rawData;
-    }
+    use RawDataTrait;
 }