+++ Classes/Domain/Repository/Event/EventRepository.php
@@ -8,21 +8,27 @@ use OliverKlee\FeUserExtraFields\Domain\Repository\DirectPersistTrait;
 use OliverKlee\Oelib\Domain\Repository\Interfaces\DirectPersist;
 use OliverKlee\Seminars\Domain\Model\Event\Event;
 use OliverKlee\Seminars\Domain\Model\Event\EventInterface;
+use OliverKlee\Seminars\Domain\Repository\AbstractRawDataCapableRepository;
 use TYPO3\CMS\Core\Database\Connection;
 use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Extbase\Persistence\Generic\Typo3QuerySettings;
 use TYPO3\CMS\Extbase\Persistence\QueryInterface;
-use TYPO3\CMS\Extbase\Persistence\Repository;
 
 /**
- * @extends Repository<Event>
+ * @extends AbstractRawDataCapableRepository<Event>
  */
-class EventRepository extends Repository implements DirectPersist
+class EventRepository extends AbstractRawDataCapableRepository implements DirectPersist
 {
     use DirectPersistTrait;
 
-    private const TABLE_NAME = 'tx_seminars_seminars';
+    /**
+     * @return non-empty-string
+     */
+    protected function getTableName(): string
+    {
+        return 'tx_seminars_seminars';
+    }
 
     /**
      * @return array<int, Event>
@@ -122,56 +128,5 @@ class EventRepository extends Repository implements DirectPersist
         $query->matching($query->equals('pid', $pageUid));
 
         return $query->execute()->toArray();
-    }
-
-    /**
-     * Sets the raw data for the provided events.
-     *
-     * This is useful e.g., for creating icons in the backend.
-     *
-     * @param array<string|int, Event> $events
-     *
-     * @internal
-     */
-    public function enrichWithRawData(array $events): void
-    {
-        if ($events === []) {
-            return;
-        }
-
-        /** @var array<positive-int, Event> $eventsByUid */
-        $eventsByUid = [];
-        foreach ($events as $event) {
-            $eventsByUid[$event->getUid()] = $event;
-        }
-
-        $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)->getQueryBuilderForTable(self::TABLE_NAME);
-        $queryBuilder->getRestrictions()->removeAll();
-        $query = $queryBuilder->select('*')
-            ->from(self::TABLE_NAME)
-            ->where(
-                $queryBuilder->expr()->in(
-                    'uid',
-                    $queryBuilder->createNamedParameter(\array_keys($eventsByUid), Connection::PARAM_INT_ARRAY)
-                )
-            );
-        if (\method_exists($query, 'executeQuery')) {
-            $queryResult = $query->executeQuery();
-        } else {
-            $queryResult = $query->execute();
-        }
-        if (\method_exists($queryResult, 'fetchAllAssociative')) {
-            $rows = $queryResult->fetchAllAssociative();
-        } else {
-            $rows = $queryResult->fetchAll();
-        }
-
-        foreach ($rows as $row) {
-            $uid = (int)$row['uid'];
-            $event = $eventsByUid[$uid] ?? null;
-            if ($event instanceof Event) {
-                $event->setRawData($row);
-            }
-        }
     }
 }